pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "qscanner"
version = "0.1.0"
description = "A qualitative stock analysis CLI tool"
readme = "README.md"
authors = [{ name = "Gemini CLI", email = "gemini-cli@google.com" }]
license = { text = "MIT" }
dependencies = [
    "typer[all]",
    "requests",
    "beautifulsoup4",
    "lxml",
    "pydantic",
    "rich",
    "google-genai"
]

[project.scripts]
qscanner = "qscanner.main:app"

[tool.setuptools.packages.find]
where = ["src"]
